requirements.txt
aiogram==3.* 
httpx 
SQLModel 
asyncpg 
python-dotenv 
APScheduler 
beautifulsoup4 
pydantic-settings
